+++ tests/conftest.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import collections.abc as c
-import gc
 import os
 from pathlib import Path
 
@@ -14,14 +13,13 @@ from flask_alembic import Alembic
 
 
 @pytest.fixture
-def app(request: pytest.FixtureRequest, tmp_path: Path) -> c.Iterator[Flask]:
+def app(request: pytest.FixtureRequest, tmp_path: Path) -> Flask:
     app = Flask(request.module.__name__, root_path=os.fspath(tmp_path))
     app.config["SQLALCHEMY_ENGINES"] = {
         "default": "sqlite://",
         "other": "sqlite://",
     }
-    yield app
-    gc.collect()
+    return app
 
 
 @pytest.fixture
@@ -32,12 +30,7 @@ def app_ctx(app: Flask) -> c.Iterator[AppContext]:
 
 @pytest.fixture
 def db(app: Flask) -> c.Iterator[SQLAlchemy]:
-    db = SQLAlchemy(app)
-    yield db
-
-    with app.app_context():
-        for engine in db.engines.values():
-            engine.dispose()
+    yield SQLAlchemy(app)
 
 
 @pytest.fixture